+++ src/components/index.js
@@ -1,0 +1,3 @@
+import Sidebar from './sidebar/sidebar';
+
+export { Sidebar };
