+++ src/exams/cs61c-sp16.js
@@ -1,8 +1,9 @@
-import React, { Component } from 'react'; import { Question } from '../components/question';
+import React, { Component } from 'react';
+import { Question } from '../components/question';
+import { Sidebar } from '../components';
 import { lx, longblank, blank, onespace, twospace, threespace, fourspace, fivespace, sixspace, sevenspace, eightspace } from '../utils';
 
 const _ = require('lodash');
-const examCode = 'cs61cfa14';
 
 const tenspace = _.repeat('&nbsp;', 10);
 const twelvespace = _.repeat('&nbsp;', 12);
@@ -726,7 +727,8 @@ class CS61CSp16 extends Component {
     scrollSpy.update();
   }
 
-  generateSidetabContainers(is_mobile) {
+  render() {
+    const examCode = 'cs61cfa14';
     const problemIDs = ['q1', 'q2', 'q3', 'q4', 'q5', 'q6'];
     const problemTitles = [
       'Q1: Instructors keep their students aligned',
@@ -736,32 +738,6 @@ class CS61CSp16 extends Component {
       'Q5: MIPS Instructions Per Second',
       'Q6: Mishmash, Hodgepodge, Potpourri'
     ];
-    return _.map(_.range(problemIDs.length), (index) => {
-      const problemID = problemIDs[index];
-      const problemTitle = problemTitles[index];
-      return (
-          <div className="sidetab-container">
-            {
-              (is_mobile) ?
-              (
-                <Link className="sidetab" to={problemID} isDynamic={true} smooth={true} duration={500}>
-                  {problemTitle}
-                </Link>
-              ) :
-              (
-                <Link activeClass="active" className="sidetab" to={problemID} spy={true} isDynamic={true} smooth={true} duration={500}>
-                  {problemTitle}
-                </Link>
-              )
-            }
-          </div>
-        );
-    });
-  }
-
-  render() {
-    const webSidetabContainers = this.generateSidetabContainers(false);
-    const mobileSidetabContainers = this.generateSidetabContainers(true);
 
     return (
       <span>
@@ -770,36 +746,7 @@ class CS61CSp16 extends Component {
         <div className="center">
           <h5>Midterm 1 | Spring 2016 | Stojanovic, Weaver</h5>
         </div>
-        <Sticky className="sidebar screen">
-          <hr className="s5" />
-          <h4>CONTENTS</h4>
-          <hr className="s2" />
-          {webSidetabContainers}
-          <hr className="s2" />
-          <h4>SOURCES</h4>
-          <hr className="s1" />
-          <div className="sidetab-container">
-            <a className="sidetab" href={process.env.PUBLIC_URL + `/exams/${examCode}-exam.pdf`} target="_blank">Exam PDF</a>
-          </div>
-          <div className="sidetab-container">
-            <a className="sidetab" href={process.env.PUBLIC_URL + `/exams/${examCode}-soln.pdf`} target="_blank">Solutions PDF</a>
-          </div>
-        </Sticky>
-        <div className="sidebar mobile">
-          <hr className="s5" />
-          <h4>CONTENTS</h4>
-          <hr className="s2" />
-          {mobileSidetabContainers}
-          <hr className="s2" />
-          <h4>SOURCES</h4>
-          <hr className="s1" />
-          <div className="sidetab-container">
-            <a className="sidetab" href={process.env.PUBLIC_URL + `/exams/${examCode}-exam.pdf`} target="_blank">Exam PDF</a>
-          </div>
-          <div className="sidetab-container">
-            <a className="sidetab" href={process.env.PUBLIC_URL + `/exams/${examCode}-soln.pdf`} target="_blank">Solutions PDF</a>
-          </div>
-        </div>
+        <Sidebar examCode={examCode} problemIDs={problemIDs} problemTitles={problemTitles} />
         <div className="content">
           <Element name="q1">
             <div className="content-spacer" />
